+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [1.10.1]
 ### Changed
-- Pascal ansistring - max length 2048
+- gdb max string length defaults to 2048
 
 ## [1.10.0] - 2025-07-30
 ### Added